+++ src/interfaces/IPushTo.ts
@@ -1,0 +1,5 @@
+import {IAssignTo} from './IAssignTo';
+
+export interface IPushTo extends IAssignTo {
+    children?: boolean;
+}